backend/requirements.txt
fastapi>=0.115.0
google-genai>=0.5.0
PyPDF2>=3.0.1
python-dotenv>=1.0.1
python-multipart>=0.0.9
uvicorn[standard]>=0.30.0
